+++ src/netspresso_trainer/models/builder.py
@@ -19,16 +19,24 @@ from .registry import (
 from .utils import load_from_checkpoint
 
 
-def load_full_model(conf_model, model_name, num_classes, model_checkpoint):
+def load_full_model(conf_model, model_name, num_classes, model_checkpoint, use_pretrained):
     model_fn: Callable[..., nn.Module] = MODEL_FULL_DICT[model_name]
+    conf_model.architecture.full.nick_name = conf_model.name # @illian01: model.name should be same with conf_model.name
     model: nn.Module = model_fn(num_classes=num_classes, conf_model_full=conf_model.architecture.full)
-    model = load_from_checkpoint(model, model_checkpoint, conf_model.load_checkpoint_head)
+    if use_pretrained:
+        model = load_from_checkpoint(
+            model, model_checkpoint,
+            load_checkpoint_head=conf_model.checkpoint.load_head,
+        )
 
     return model
 
 
 def load_backbone_and_head_model(
-        conf_model, task, backbone_name, head_name, num_classes, model_checkpoint, img_size, freeze_backbone):
+        conf_model, task, backbone_name, head_name, num_classes,
+        model_checkpoint, use_pretrained, freeze_backbone,
+        img_size
+    ):
     if task not in TASK_MODEL_DICT:
         raise ValueError(
             f"No such task(s) named: {task}. This should be included in SUPPORTING_TASK_LIST ({SUPPORTING_TASK_LIST})")
@@ -59,18 +67,31 @@ def load_backbone_and_head_model(
 
     # Assemble model and load checkpoint
     model = TASK_MODEL_DICT[task](conf_model, backbone, neck, head, freeze_backbone)
-    model = load_from_checkpoint(model, model_checkpoint, conf_model.load_checkpoint_head)
+    if use_pretrained:
+        model = load_from_checkpoint(
+            model, model_checkpoint,
+            load_checkpoint_head=conf_model.checkpoint.load_head,
+        )
     return model
 
 
-def build_model(conf_model, task, num_classes, model_checkpoint, img_size) -> nn.Module:
+def build_model(conf_model, task, num_classes, model_checkpoint, use_pretrained, img_size) -> nn.Module:
 
     if conf_model.single_task_model:
         model_name = str(conf_model.architecture.full.name).lower()
-        return load_full_model(conf_model, model_name, num_classes, model_checkpoint)
+        return load_full_model(
+            conf_model, model_name, num_classes,
+            model_checkpoint=model_checkpoint,
+            use_pretrained=use_pretrained
+        )
 
     backbone_name = str(conf_model.architecture.backbone.name).lower()
     head_name = str(conf_model.architecture.head.name).lower()
     freeze_backbone = conf_model.freeze_backbone
     return load_backbone_and_head_model(
-        conf_model, task, backbone_name, head_name, num_classes, model_checkpoint, img_size, freeze_backbone)
+        conf_model, task, backbone_name, head_name, num_classes,
+        model_checkpoint=model_checkpoint,
+        use_pretrained=use_pretrained,
+        freeze_backbone=freeze_backbone,
+        img_size=img_size
+    )