+++ src/netspresso_trainer/models/full/experimental/pidnet.py
@@ -16,12 +16,14 @@ use_align_corners = False
 
 
 class PIDNet(nn.Module):
-
+    task = 'segmentation'
     def __init__(
-            self, 
-            params: Optional[Dict] = None
+        self, 
+        params: Optional[Dict] = None
     ) -> None:
         super(PIDNet, self).__init__()
+        self.name = params.nick_name
+        
         num_classes = params.num_classes
         m = params.m
         n = params.n
@@ -36,10 +38,10 @@ class PIDNet(nn.Module):
         self.conv1 = nn.Sequential(
             ConvLayer(in_channels=3, out_channels=planes,
                       kernel_size=3, stride=2, padding=1,
-                      norm_type='batch_norm', act_type='relu'),
+                      norm_type='batch_norm', act_type='relu', bias=True),
             ConvLayer(in_channels=planes, out_channels=planes,
                       kernel_size=3, stride=2, padding=1,
-                      norm_type='batch_norm', act_type='relu')
+                      norm_type='batch_norm', act_type='relu', bias=True)
         )
 
         self.relu = nn.ReLU(inplace=True)
@@ -140,6 +142,10 @@ class PIDNet(nn.Module):
         layer = block(inplanes, planes, stride, downsample, expansion=expansion, no_out_act=True)
 
         return layer
+
+    @property
+    def head_list(self):
+        return ('seghead_p', 'seghead_d', 'final_layer')
 
     @property
     def device(self):